Improve local dev server
