+++ cli/package.json
@@ -33,6 +33,7 @@
     "@oclif/core": "^4",
     "chalk": "^5.3.0",
     "chokidar": "^4.0.1",
+    "express": "^4.21.1",
     "gradient-string": "^3.0.0",
     "open": "^10.1.0",
     "ora": "^8.1.0",
@@ -43,6 +44,7 @@
     "@eslint/js": "^9.14.0",
     "@oclif/test": "^4",
     "@types/eslint__js": "^8.42.3",
+    "@types/express": "^5.0.0",
     "@types/node": "^22",
     "@types/picomatch": "^3.0.1",
     "@types/semver": "^7.5.8",
